+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased] - ReleaseDate
 
+## [0.1.4] - 2022-09-15
+
 ### Added
 
 * `--check`: show diff if README is not updated
@@ -46,7 +48,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * First release
 
 <!-- next-url -->
-[Unreleased]: https://github.com/gifnksm/cargo-sync-rdme/compare/v0.1.3...HEAD
+[Unreleased]: https://github.com/gifnksm/cargo-sync-rdme/compare/v0.1.4...HEAD
+[0.1.4]: https://github.com/gifnksm/cargo-sync-rdme/compare/v0.1.3...v0.1.4
 [0.1.3]: https://github.com/gifnksm/cargo-sync-rdme/compare/v0.1.2...v0.1.3
 [0.1.2]: https://github.com/gifnksm/cargo-sync-rdme/compare/v0.1.1...v0.1.2
 [0.1.1]: https://github.com/gifnksm/cargo-sync-rdme/compare/v0.1.0...v0.1.1